+++ src/components/timeline/index.ts
@@ -1,0 +1,14 @@
+export { Timeline, TimelineData, TimelineFooter, TimelineItem, TimelineTitle } from "./Timeline";
+export type {
+  TimelineClassNames,
+  TimelineDataClassNames,
+  TimelineDataProps,
+  TimelineFooterClassNames,
+  TimelineFooterProps,
+  TimelineItemClassNames,
+  TimelineItemProps,
+  TimelineItemStatus,
+  TimelineProps,
+  TimelineTitleClassNames,
+  TimelineTitleProps,
+} from "../../types/timeline";