+++ src/index.ts
@@ -41,6 +41,7 @@ export * from "./components/sheet";
 export * from "./components/sidebar";
 export * from "./components/sortable";
 export * from "./components/tabs";
+export * from "./components/timeline";
 export * from "./components/textarea";
 export * from "./components/toast";
 export * from "./components/tooltip";